+++ www/wp-content/themes/hittheroad-2021/woocommerce/checkout/form-coupon.php
@@ -12,7 +12,7 @@
  *
  * @see https://docs.woocommerce.com/document/template-structure/
  * @package WooCommerce\Templates
- * @version 7.0.1
+ * @version 9.8.0
  */
 
 defined('ABSPATH') || exit;
@@ -26,15 +26,15 @@ if (isset($notification) && $notification !== '') : ?>
 <?php endif ?>
 
 <div class="woocommerce-form-coupon-toggle">
-	<?php wc_print_notice(apply_filters('woocommerce_checkout_coupon_message', esc_html__('Have a coupon?', 'woocommerce') . ' <a href="#" class="showcoupon">' . esc_html__('Click here to enter your code', 'woocommerce') . '</a>'), 'notice'); ?>
+	<?php wc_print_notice(apply_filters('woocommerce_checkout_coupon_message', esc_html__('Have a coupon?', 'woocommerce') . ' <a href="#" role="button" aria-label="' . esc_attr__('Enter your coupon code', 'woocommerce') . '" aria-controls="woocommerce-checkout-form-coupon" aria-expanded="false" class="showcoupon">' . esc_html__('Click here to enter your code', 'woocommerce') . '</a>'), 'notice'); ?>
 </div>
 
-<form class="checkout_coupon woocommerce-form-coupon" method="post" style="display:none">
+<form class="checkout_coupon woocommerce-form-coupon" method="post" style="display:none" id="woocommerce-checkout-form-coupon">
 	<p><?php esc_html_e('If you have a coupon code, please apply it below.', 'woocommerce'); ?></p>
 
 	<div class="d-inline-flex justify-content-start">
 		<label for="coupon_code" class="screen-reader-text"><?php esc_html_e('Coupon:', 'woocommerce'); ?></label>
-		<input type="text" name="coupon_code" class="form-control me-1" placeholder="<?php esc_attr_e('Coupon code', 'woocommerce'); ?>" id="coupon_code" arial-label="<?php esc_attr_e('Coupon code', 'woocommerce'); ?>">
-		<button type="submit" class="btn btn-primary btn-outline" name="apply_coupon">Appliquer</button>
+		<input type="text" name="coupon_code" class="form-control me-1" placeholder="<?php esc_attr_e('Coupon code', 'woocommerce'); ?>" id="coupon_code" arial-label="<?php esc_attr_e('Coupon code', 'woocommerce'); ?>" value="">
+		<button type="submit" class="btn btn-primary btn-outline" name="apply_coupon" value="<?php esc_attr_e('Apply coupon', 'woocommerce'); ?>">Appliquer</button>
 	</div>
 </form>