+++ www/wp-content/themes/hittheroad-2021/woocommerce/global/quantity-input.php
@@ -12,7 +12,7 @@
  *
  * @see     https://docs.woocommerce.com/document/template-structure/
  * @package WooCommerce\Templates
- * @version 9.4.0
+ * @version 10.1.0
  *
  * @var bool   $readonly If the input should be set to readonly mode.
  * @var string $type     The input type attribute.
@@ -45,7 +45,9 @@ $label = !empty($args['product_name']) ? sprintf(esc_html__('%s quantity', 'wooc
 			size="4"
 		<?php endif; ?>
 		min="<?= esc_attr($min_value); ?>"
-		max="<?= esc_attr(0 < $max_value ? $max_value : ''); ?>"
+		<?php if (0 < $max_value) : ?>
+			max="<?php echo esc_attr( $max_value ); ?>"
+		<?php endif; ?>
 		<?php if (!$readonly) : ?>
 			step="<?= esc_attr($step); ?>"
 			placeholder="<?= esc_attr($placeholder); ?>"