+++ www/wp-content/themes/hittheroad-2021/woocommerce/notices/notice.php
@@ -12,7 +12,7 @@
  *
  * @see     https://docs.woocommerce.com/document/template-structure/
  * @package WooCommerce\Templates
- * @version 8.6.0
+ * @version 10.2.0
  */
 
 if (!defined('ABSPATH')) {
@@ -25,7 +25,7 @@ if (!$notices) {
 ?>
 
 <?php foreach ($notices as $notice) : ?>
-	<div class="alert alert-info" role="alert"<?= wc_get_notice_data_attr($notice); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped  ?>>
+	<div class="alert alert-info" role="status" <?= wc_get_notice_data_attr($notice); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped  ?>>
 		<?= wc_kses_notice($notice['notice']); ?>
 	</div>
 <?php endforeach; ?>